+++ demo/components/index.ts
@@ -1,0 +1,3 @@
+export { SimpleDemoComponent } from './simple.component';
+export { CustomKeyDemoComponent } from './custom-key.component';
+export { CustomTemplateComponent } from './custom-template';