+++ demo/demo.module.ts
@@ -2,9 +2,19 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { NgxDualListboxModule } from '../src';
 import { DemoComponent } from './demo.component';
+import {
+  SimpleDemoComponent,
+  CustomKeyDemoComponent,
+  CustomTemplateComponent
+} from './components';
 
 @NgModule({
-  declarations: [DemoComponent],
+  declarations: [
+    DemoComponent,
+    SimpleDemoComponent,
+    CustomKeyDemoComponent,
+    CustomTemplateComponent
+  ],
   imports: [BrowserModule, NgxDualListboxModule.forRoot()],
   bootstrap: [DemoComponent]
 })